requirements.txt
pandas
Pillow
pyperclip
pywin32
selenium==3.14.1
webdriver-manager==3.5.2